# doc/swagger.yml
openapi: 3.0.0
info:
  description: "This API handles a collection of movies and series for easy ordering along locations, type, categories."
  version: "xxx"
  title: "Porygon API"
  contact:
    name: "Aurélien Hugues"
    email: "me@aurelienhugues.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: http://localhost:4000/api/v1
    description: Dev server
  - url: https://api.porygon.aurelienhugues.com
    description: Production server
components:
  schemas:
...
    StatusMessage:
      properties:
        code:
          type: integer
        userMessage:
          type: string
      required:
        - code
        - userMessage
    OutputCategory:
      properties:
        id:
          type: integer
        label:
          type: string
        description:
          type: string
      required:
        - id
        - label
        - description
    InputCategory:
      properties:
        label:
          type: string
        description:
          type: string
      required:
        - label
    OutputLocation:
      properties:
        id:
          type: integer
        location:
          type: string
        is_physical:
          description: True if the location is physical, False if digital
          type: boolean
      required:
        - id
        - location
        - is_physical
    InputLocation:
      properties:
        location:
          type: string
        is_physical:
          description: True if the location is physical, False if digital
          type: boolean
      required:
        - location
        - is_physical
...
    FullStats:
      properties:
        movie_count:
          type: integer
        serie_count:
          type: integer
        location_count:
          type: integer
        category_count:
          type: integer
      required:
        - movie_count
        - serie_count
        - location_count
        - category_count
...
paths:
...
  /categories:
    get:
      summary: Get list of categories
      description: Get the full list of categories with optional query parameters to limit the search
      tags:
        - categories
      parameters:
        - in: query
          name: 'attributes'
          description: List of attributes to return, separated by a comma
          schema:
            type: array
            items:
              enum:
                - label
                - description
                - id
              type: string
          style: form
          explode: false
        - in: query
          name: label
          description: Search parameter for the label
          schema:
            type: string
        - in: query
          name: sort
          description: Order to use for sorting
          schema:
            type: string
          examples:
            label:
              value: +label
              summary: Ascending order along label
            description:
              value: -description
              summary: Descending order along description
      responses:
        '200':
          description: A JSON array of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutputCategory'
        '400':
          description: Invalid field
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
              example:
                code: 400
                userMessage: "Invalid field 'toto' in query"
        '500':
          description: Unexpected error
    post:
      summary: Create a new category
      tags:
        - categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputCategory'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
              example:
                code: 201
                userMessage: 'Category successfully created'
        '400':
          description: Invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
              example:
                code: 400
                userMessage: "Invalid field 'toto' in query"
            text/html:
              schema:
                type: string
                example: |
                  <!DOCTYPE html>
                    <html>
                    
                      <head>
                        <title></title>
                        <link rel="stylesheet" href="/stylesheets/style.css">
                      </head>
                      
                      <body>
                        <h1>Unauthorized field 'toto' in query</h1>
                        <h2>400</h2>
                        <pre></pre>
                      </body>
                  
                  </html>
        '500':
          description: Unexpected error
...
  /locations:
    get:
      summary: Get list of locations
      description: Get the full list of locations
      tags:
        - locations
      responses:
        '200':
          description: A JSON array of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutputLocation'
        '500':
          description: Unexpected error
    post:
      summary: Create a new location
      tags:
        - locations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputLocation'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
              example:
                code: 201
                userMessage: 'Location successfully created'
        '400':
          description: Invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
              example:
                code: 400
                userMessage: "Invalid field 'toto' in query"
            text/html:
              schema:
                type: string
                example: |
                  <!DOCTYPE html>
                    <html>
                    
                      <head>
                        <title></title>
                        <link rel="stylesheet" href="/stylesheets/style.css">
                      </head>
                      
                      <body>
                        <h1>Unauthorized field 'toto' in query</h1>
                        <h2>400</h2>
                        <pre></pre>
                      </body>
                  
                  </html>
        '500':
          description: Unexpected error
...
  /stats:
    get:
      summary: Get complete stats for the database
      description: Get the number of all elements in the database
      tags:
        - stats
      responses:
        '200':
          description: Complete database stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullStats'
        '500':
          description: Unexpected error
...